refactor: extract issue search/bulk to github/issue_search.py (#140)
